# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PoolRide {
  id: number;
  fromAddress: string;
  toAddress: string;
  fromLat: number;
  fromLng: number;
  toLat: number;
  toLng: number;
  date: string;
  time: string;
  seats: number;
  seatsLeft: number;
  status: string;
  driverName: string;
  /** @nullable */
  driverAvatar: string | null;
  driverRating: number;
  driverTrips: number;
  farePerSeat: number;
  luggageAllowed?: boolean;
  smokingAllowed?: boolean;
  musicAllowed?: boolean;
  /** @nullable */
  genderPreference?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  profileId?: number | null;
  createdAt: string;
}

export interface PoolRideInput {
  fromAddress: string;
  toAddress: string;
  fromLat: number;
  fromLng: number;
  toLat: number;
  toLng: number;
  date: string;
  time: string;
  seats: number;
  farePerSeat: number;
  driverName: string;
  driverAvatar?: string;
  luggageAllowed?: boolean;
  smokingAllowed?: boolean;
  musicAllowed?: boolean;
  genderPreference?: string;
  notes?: string;
  profileId?: number;
}

export interface PoolRideUpdate {
  status?: string;
  seatsLeft?: number;
  notes?: string;
  farePerSeat?: number;
}

export interface JoinRequest {
  id: number;
  rideId: number;
  passengerName: string;
  /** @nullable */
  passengerAvatar?: string | null;
  seats: number;
  /** @nullable */
  pickupAddress?: string | null;
  status: string;
  /** @nullable */
  profileId?: number | null;
  createdAt: string;
}

export interface JoinRideInput {
  passengerName: string;
  passengerAvatar?: string;
  seats: number;
  pickupAddress?: string;
  profileId?: number;
}

export interface JoinRequestUpdate {
  status: string;
}

export interface UserProfile {
  id: number;
  name: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  verified?: boolean;
  trustScore?: number;
  totalRides?: number;
  rating?: number;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface UserProfileInput {
  name: string;
  avatar?: string;
  phone?: string;
  email?: string;
  bio?: string;
}

export interface UserProfileUpdate {
  name?: string;
  avatar?: string;
  phone?: string;
  email?: string;
  bio?: string;
  verified?: boolean;
}

export interface Review {
  id: number;
  profileId: number;
  reviewerName: string;
  /** @nullable */
  reviewerAvatar?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  profileId: number;
  reviewerName: string;
  reviewerAvatar?: string;
  rating: number;
  comment?: string;
}

export interface Message {
  id: number;
  rideId: number;
  senderName: string;
  /** @nullable */
  senderAvatar?: string | null;
  text: string;
  createdAt: string;
}

export interface MessageInput {
  rideId: number;
  senderName: string;
  senderAvatar?: string;
  text: string;
}

export interface TopRoute {
  from: string;
  to: string;
  count: number;
}

export interface RideStats {
  totalRides: number;
  activeRides: number;
  totalPassengers: number;
  citiesServed: number;
  avgRating: number;
  ridesThisWeek?: number;
  topRoutes?: TopRoute[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export type ListRidesParams = {
fromLat?: number;
fromLng?: number;
toLat?: number;
toLng?: number;
date?: string;
seats?: number;
status?: string;
};

export type ListReviewsParams = {
profileId?: number;
};

export type ListMessagesParams = {
rideId: number;
};