# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}