# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  rideId: number;
  senderName: string;
  /** @nullable */
  senderAvatar?: string | null;
  text: string;
  createdAt: string;
}