# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  rideId: number;
  senderName: string;
  senderAvatar?: string;
  text: string;
}