# lib/api-zod/src/generated/types/poolRide.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */

export interface PoolRide {
  id: number;
  fromAddress: string;
  toAddress: string;
  fromLat: number;
  fromLng: number;
  toLat: number;
  toLng: number;
  date: string;
  time: string;
  seats: number;
  seatsLeft: number;
  status: string;
  driverName: string;
  /** @nullable */
  driverAvatar: string | null;
  driverRating: number;
  driverTrips: number;
  farePerSeat: number;
  luggageAllowed?: boolean;
  smokingAllowed?: boolean;
  musicAllowed?: boolean;
  /** @nullable */
  genderPreference?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  profileId?: number | null;
  createdAt: string;
}