# lib/api-zod/src/generated/types/poolRideUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */

export interface PoolRideUpdate {
  status?: string;
  seatsLeft?: number;
  notes?: string;
  farePerSeat?: number;
}