# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  profileId: number;
  reviewerName: string;
  /** @nullable */
  reviewerAvatar?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}