# lib/api-zod/src/generated/types/rideStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CabScan API — ride comparison and pool rides
 * OpenAPI spec version: 0.1.0
 */
import type { TopRoute } from './topRoute';

export interface RideStats {
  totalRides: number;
  activeRides: number;
  totalPassengers: number;
  citiesServed: number;
  avgRating: number;
  ridesThisWeek?: number;
  topRoutes?: TopRoute[];
}